+++ src/generate/generic.rs
@@ -53,7 +53,7 @@ pub trait RegisterSpec {
 /// Raw field type
 pub trait FieldSpec: Sized {
     /// Raw field type (`u8`, `u16`, `u32`, ...).
-    type Ux: Copy + PartialEq + From<Self>;
+    type Ux: Copy + core::fmt::Debug + PartialEq + From<Self>;
 }
 
 /// Marker for fields with fixed values
@@ -258,6 +258,15 @@ impl<REG: Readable + Writable> Reg<REG> {
     }
 }
 
+impl<REG: Readable> core::fmt::Debug for crate::generic::Reg<REG>
+where
+    R<REG>: core::fmt::Debug
+{
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        core::fmt::Debug::fmt(&self.read(), f)
+    }
+}
+
 #[doc(hidden)]
 pub mod raw {
     use super::{marker, BitM, FieldSpec, RegisterSpec, Unsafe, Writable};
@@ -433,6 +442,12 @@ impl<FI: FieldSpec> FieldReader<FI> {
     }
 }
 
+impl<FI: FieldSpec> core::fmt::Debug for FieldReader<FI> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        core::fmt::Debug::fmt(&self.bits, f)
+    }
+}
+
 impl<FI> PartialEq<FI> for FieldReader<FI>
 where
     FI: FieldSpec + Copy,
@@ -469,6 +484,12 @@ impl<FI> BitReader<FI> {
     #[inline(always)]
     pub const fn bit_is_set(&self) -> bool {
         self.bit()
+    }
+}
+
+impl<FI> core::fmt::Debug for BitReader<FI> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        core::fmt::Debug::fmt(&self.bits, f)
     }
 }